#3130: added markus component
git-svn-id: https://ncisvn.nci.nih.gov/svn/geworkbench/trunk/geworkbench-web@9773 8b29bb6e-589a-435a-9a12-3ab907be3462

+++ src/org/geworkbenchweb/dataset/UDataSetUpload.java
@@ -15,7 +15,7 @@ import com.vaadin.ui.Window;
 @SuppressWarnings("serial")
 public class UDataSetUpload extends Window {
     
-    private static final String[] files		= 	new String[] { "Expression File" };
+    private static final String[] files		= 	new String[] { "Expression File", "PDB File" };
     private static final String initialText = 	"Enter description here.";
    
     private VerticalLayout dataLayout; 
@@ -109,7 +109,8 @@ public class UDataSetUpload extends Window {
     	
     	parseInit(dataFile, null, fileType, dataDescription);
     		
-    	dataFile.delete();
+    	if (!fileType.equals("PDB File"))
+    		dataFile.delete();
     	
     }
